refactor: ai system
