+++ liubai-backends/liubai-laf/cloud-functions/ai-shared.ts
@@ -1,0 +1,289 @@
+import cloud from "@lafjs/cloud"
+import type { 
+  LiuAi, 
+  AiBot,
+  AiEntry,
+  AiCharacter,
+  Wx_Gzh_Send_Msg,
+  Wx_Gzh_Send_Msgmenu,
+  Wx_Gzh_Send_Msgmenu_Item,
+} from "@/common-types"
+import { WxGzhSender } from "@/service-send"
+import { 
+  checkAndGetWxGzhAccessToken,
+  MarkdownParser,
+} from "@/common-util"
+import { aiBots } from "@/ai-prompt"
+import { useI18n, aiLang } from "@/common-i18n"
+import { WxGzhUploader } from "@/file-utils"
+
+const db = cloud.database()
+const _ = db.command
+
+export class AiShared {
+
+  static getApiEndpointFromBot(
+    bot: AiBot
+  ): LiuAi.ApiEndpoint | undefined {
+    const _env = process.env
+    const p = bot.provider
+    const p2 = bot.secondaryProvider
+
+    let apiKey: string | undefined
+    let baseURL: string | undefined
+    let defaultHeaders = bot.metaData?.defaultHeaders
+
+    // If secondaryProvider exists, use it first
+    if(p2 === "siliconflow") {
+      apiKey = _env.LIU_SILICONFLOW_API_KEY
+      baseURL = _env.LIU_SILICONFLOW_BASE_URL
+    }
+    else if(p2 === "gitee-ai") {
+      apiKey = _env.LIU_GITEE_AI_API_KEY
+      baseURL = _env.LIU_GITEE_AI_BASE_URL
+    }
+    else if(p === "baichuan") {
+      apiKey = _env.LIU_BAICHUAN_API_KEY
+      baseURL = _env.LIU_BAICHUAN_BASE_URL
+    }
+    else if(p === "deepseek") {
+      apiKey = _env.LIU_DEEPSEEK_API_KEY
+      baseURL = _env.LIU_DEEPSEEK_BASE_URL
+    }
+    else if(p === "minimax") {
+      apiKey = _env.LIU_MINIMAX_API_KEY
+      baseURL = _env.LIU_MINIMAX_BASE_URL
+    }
+    else if(p === "moonshot") {
+      apiKey = _env.LIU_MOONSHOT_API_KEY
+      baseURL = _env.LIU_MOONSHOT_BASE_URL
+    }
+    else if(p === "stepfun") {
+      apiKey = _env.LIU_STEPFUN_API_KEY
+      baseURL = _env.LIU_STEPFUN_BASE_URL
+    }
+    else if(p === "zero-one") {
+      apiKey = _env.LIU_YI_API_KEY
+      baseURL = _env.LIU_YI_BASE_URL
+    }
+    else if(p === "zhipu") {
+      apiKey = _env.LIU_ZHIPU_API_KEY
+      baseURL = _env.LIU_ZHIPU_BASE_URL
+    }
+    
+    if(apiKey && baseURL) {
+      return { apiKey, baseURL, defaultHeaders }
+    }
+  }
+
+  static getCharacterName(character?: AiCharacter) {
+    if(!character) return
+    let name = ""
+    const bot = aiBots.find(v => v.character === character)
+    if(bot) name = bot.name
+    return name
+  }
+
+  static getGzhType() {
+    const _env = process.env
+    return _env.LIU_WX_GZ_TYPE ?? "subscription_account"
+  }
+
+
+}
+
+export class TellUser {
+
+  static async text(
+    entry: AiEntry, 
+    text: string,
+    from?: AiBot,
+    fromCharacter?: AiCharacter
+  ) {
+    const { wx_gzh_openid } = entry
+
+    // 1. send to wx gzh
+    if(wx_gzh_openid) {
+      // console.warn("markdown: ")
+      // console.log(text)
+      text = MarkdownParser.mdToWxGzhText(text)
+      // console.warn("wx gzh text: ")
+      // console.log(text)
+
+      const obj1: Wx_Gzh_Send_Msg = {
+        msgtype: "text",
+        text: { content: text },
+      }
+      this._fillWxGzhKf(obj1, from, fromCharacter)
+      const res1 = await this._sendToWxGzh(wx_gzh_openid, obj1)
+      return res1
+    }
+
+  }
+
+  static async image(
+    entry: AiEntry,
+    imageUrl: string,
+    from?: AiBot,
+    fromCharacter?: AiCharacter,
+  ) {
+    const { wx_gzh_openid } = entry
+
+    // 1. send to wx gzh
+    if(wx_gzh_openid) {
+      const res1 = await WxGzhUploader.mediaByUrl(imageUrl)
+      const media_id = res1?.media_id
+      if(!media_id) return
+
+      const obj2: Wx_Gzh_Send_Msg = {
+        msgtype: "image",
+        image: { media_id },
+      }
+      this._fillWxGzhKf(obj2, from, fromCharacter)
+      const res2 = await this._sendToWxGzh(wx_gzh_openid, obj2)
+      return res2
+    }
+  }
+
+
+  static async menu(
+    entry: AiEntry,
+    prefixMessage: string,
+    menuList: LiuAi.MenuItem[],
+    suffixMessage: string,
+    fromCharacter?: AiCharacter
+  ) {
+    const _env = process.env
+    const gzhType = AiShared.getGzhType()
+    const { wx_gzh_openid, user } = entry
+    const { t } = useI18n(aiLang, { user })
+
+    // 1. localize the menuList
+    const wx_menu_list: Wx_Gzh_Send_Msgmenu_Item[] = []
+    for(let i=0; i<menuList.length; i++) {
+      const v = menuList[i]
+      const { operation, character } = v
+
+      if(operation === "clear_history") {
+        wx_menu_list.push({ id: "clear_history", content: t("clear_context") })
+        continue
+      }
+
+      if(operation === "kick" && character) {
+        const characterName = AiShared.getCharacterName(character)
+        if(!characterName) continue
+        wx_menu_list.push({ id: "kick_" + character, content: t("kick") + characterName })
+      }
+
+      if(operation === "add" && character) {
+        const characterName = AiShared.getCharacterName(character)
+        if(!characterName) continue
+        wx_menu_list.push({ id: "add_" + character, content: t("add") + characterName })
+      }
+
+      if(operation === "continue" && character) {
+        const characterName = AiShared.getCharacterName(character)
+        if(!characterName) continue
+        wx_menu_list.push({
+          id: "continue_" + character,
+          content: t("continue_bot", { botName: characterName })
+        })
+
+        // turn markdown to plain-text for wx gzh
+        if(wx_gzh_openid) {
+          prefixMessage = MarkdownParser.mdToWxGzhText(prefixMessage)
+        }
+      }
+
+    }
+
+    // 2. send to wx gzh
+    if(wx_gzh_openid) {
+      if(gzhType === "subscription_account") {
+        console.warn("we cannot send the menu to the user due to subscription_account")
+        return
+      }
+
+      const obj2: Wx_Gzh_Send_Msgmenu = {
+        msgtype: "msgmenu",
+        msgmenu: {
+          head_content: prefixMessage,
+          list: wx_menu_list,
+          tail_content: suffixMessage,
+        }
+      }
+      this._fillWxGzhKf(obj2, undefined, fromCharacter)
+      const res2 = await this._sendToWxGzh(wx_gzh_openid, obj2)
+      return res2
+    }
+    
+
+  }
+
+  static async typing(entry: AiEntry) {
+    const { wx_gzh_openid } = entry
+
+    // 1. to wx gzh
+    if(wx_gzh_openid) {
+      const wxGzhAccessToken = await checkAndGetWxGzhAccessToken()
+      if(!wxGzhAccessToken) return
+      WxGzhSender.sendTyping(wx_gzh_openid, wxGzhAccessToken)
+    }
+  }
+
+  private static _fillWxGzhKf(
+    obj: Wx_Gzh_Send_Msg,
+    bot?: AiBot,
+    character?: AiCharacter,
+  ) {
+    const kf_account = this._getWxGzhKfAccount(bot, character)
+    if(kf_account) {
+      obj.customservice = { kf_account }
+    }
+  }
+
+  private static _getWxGzhKfAccount(
+    bot?: AiBot,
+    character?: AiCharacter,
+  ) {
+    let c = bot?.character ?? character
+    if(!c) return
+
+    const _env = process.env
+    if(c === "baixiaoying") {
+      return _env.LIU_WXGZH_KF_BAIXIAOYING
+    }
+    else if(c === "deepseek") {
+      return _env.LIU_WXGZH_KF_DEEPSEEK
+    }
+    else if(c === "ds-reasoner") {
+      return _env.LIU_WXGZH_KF_DS_REASONER
+    }
+    else if(c === "hailuo") {
+      return _env.LIU_WXGZH_KF_HAILUO
+    }
+    else if(c === "kimi") {
+      return _env.LIU_WXGZH_KF_KIMI
+    }
+    else if(c === "wanzhi") {
+      return _env.LIU_WXGZH_KF_WANZHI
+    }
+    else if(c === "yuewen") {
+      return _env.LIU_WXGZH_KF_YUEWEN
+    }
+    else if(c === "zhipu") {
+      return _env.LIU_WXGZH_KF_ZHIPU
+    }
+  }
+
+  private static async _sendToWxGzh(
+    wx_gzh_openid: string,
+    obj: Wx_Gzh_Send_Msg,
+  ) {
+    const accessToken = await checkAndGetWxGzhAccessToken()
+    if(!accessToken) return
+    const res = await WxGzhSender.sendMessage(wx_gzh_openid, accessToken, obj)
+    return res
+  }
+
+}
